# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Custom Analytic - WW2 API",
        "description": "API Documentation for Custom Analytic - WW2",
        "contact": {
            "email": "flazen.edu@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/aircraft": {
            "get": {
                "tags": [
                    "Aircraft"
                ],
                "summary": "Show all aircraft module or combined API (all data & stats)",
                "operationId": "2a66a770ca931b589ad3756a4b983315",
                "parameters": [
                    {
                        "name": "limit_data_all",
                        "in": "query",
                        "description": "Limit the number of aircraft to show",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    },
                    {
                        "name": "order_data_all",
                        "in": "query",
                        "description": "Order the aircraft by ascending or descending",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "asc",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    },
                    {
                        "name": "search_data_all",
                        "in": "query",
                        "description": "Search term for filtering aircraft",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "%20"
                        }
                    },
                    {
                        "name": "limit_stats_by_role",
                        "in": "query",
                        "description": "Limit the number of role to show",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 7
                        }
                    },
                    {
                        "name": "limit_stats_by_country",
                        "in": "query",
                        "description": "Limit the number of country to show",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 7
                        }
                    },
                    {
                        "name": "limit_stats_by_manufacturer",
                        "in": "query",
                        "description": "Limit the number of manufacturer to show",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 7
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "aircraft module found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "post": {
                "tags": [
                    "Aircraft"
                ],
                "summary": "Add aircraft",
                "operationId": "8c7f9d6a0c59a9391edb65788acdb44e",
                "responses": {
                    "200": {
                        "description": "New aircraft ... has been created"
                    },
                    "409": {
                        "description": "Data is already exist"
                    },
                    "422": {
                        "description": "{validation_msg}"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/aircraft/limit/{limit}/order/{order}/find/{search}": {
            "get": {
                "tags": [
                    "Aircraft"
                ],
                "summary": "Show all aircraft with pagination, ordering, and search",
                "operationId": "e35ab7a5e47baf72a239e43adcf891c9",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "path",
                        "description": "Number of aircraft per page",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    },
                    {
                        "name": "order",
                        "in": "path",
                        "description": "Order by field name",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "asc"
                        }
                    },
                    {
                        "name": "search",
                        "in": "path",
                        "description": "Search term based on the field name or manufacturer",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Ar"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "aircraft found"
                    },
                    "404": {
                        "description": "aircraft not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/aircraft/summary": {
            "get": {
                "tags": [
                    "Aircraft"
                ],
                "summary": "Show aircraft summary",
                "operationId": "3655b06802b4a374f6f37688ad4bb83d",
                "responses": {
                    "200": {
                        "description": "aircraft found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/aircraft/total/byrole/{limit}": {
            "get": {
                "tags": [
                    "Aircraft"
                ],
                "summary": "Total aircraft by role",
                "operationId": "86721f354fdaa5c6c4fe931a692510cb",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "path",
                        "description": "Number of aircraft per page",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "aircraft found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/aircraft/total/bymanufacturer/{limit}": {
            "get": {
                "tags": [
                    "Aircraft"
                ],
                "summary": "Total aircraft by manufacturer",
                "operationId": "c8cdf67e6eb3116f51d8b298b097db5c",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "path",
                        "description": "Number of aircraft per page",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "aircraft found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/aircraft/total/bysides": {
            "get": {
                "tags": [
                    "Aircraft"
                ],
                "summary": "Total aircraft by sides",
                "operationId": "69d787263b94e0c1244e5dde786a4891",
                "responses": {
                    "200": {
                        "description": "aircraft found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/aircraft/total/bycountry/{limit}": {
            "get": {
                "tags": [
                    "Aircraft"
                ],
                "summary": "Total aircraft by country",
                "operationId": "8438062cc4073c0cef01044617e144ca",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "path",
                        "description": "Number of aircraft per page",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "aircraft found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
    },
...
}